Rename timingwheel to tickwheel.
