+++ tickwheel/tickWheel.go
@@ -1,4 +1,4 @@
-package timingwheel
+package tickwheel
 
 import (
 	"fmt"
